py/requirements.txt
openpyxl>=3.1,<4
pandas>=2.2,<3
gspread>=6,<7
google-auth>=2,<3
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0
numpy
